+++ src/collective/documentgenerator/tests/test_base_helper_view.py
@@ -3,33 +3,41 @@
 import appy
 from StringIO import StringIO
 from plone import api
-from collective.documentgenerator.demo.helper import DemoHelperView
-from collective.documentgenerator.testing import ArchetypesIntegrationTests
+from collective.documentgenerator.config import HAS_PLONE_5
+from collective.documentgenerator.demo.helper import BaseDemoHelperView
+from collective.documentgenerator.demo.helper import ATDemoHelperView
+from collective.documentgenerator.demo.helper import DXDemoHelperView
+from collective.documentgenerator.testing import DexterityIntegrationTests
 
 
-class TestBaseHelperViewMethods(ArchetypesIntegrationTests):
+class TestBaseHelperViewMethods(DexterityIntegrationTests):
     """
     Test base helper view's methods.
     """
 
     def setUp(self):
         super(TestBaseHelperViewMethods, self).setUp()
-        self.view = self.AT_topic.unrestrictedTraverse('@@document_generation_helper_view')
+        self.view = self.content.unrestrictedTraverse(
+            '@@document_generation_helper_view')
 
     def test_translate_method(self):
         msgid = u'month_may'
         translation = self.view.translate(msgid, domain='plonelocales')
-        self.assertTrue(translation == u'May')
+        self.assertTrue(translation == u'Mai')
 
     def test_getDGHV(self):
         new_dghv = self.view.getDGHV(self.portal['podtemplates'])
-        isinstance(new_dghv, DemoHelperView)
+        self.assertTrue(isinstance(new_dghv, BaseDemoHelperView))
+        if HAS_PLONE_5:
+            self.assertTrue(isinstance(new_dghv, DXDemoHelperView))
+        else:
+            self.assertTrue(isinstance(new_dghv, ATDemoHelperView))
         self.assertEqual(new_dghv.real_context, self.portal['podtemplates'])
         self.assertEqual(new_dghv.display('title'), 'POD Templates')
 
     def test_get_state(self):
-        # no workflow
-        self.assertEqual(self.view.get_state(title=True), '-')
+        # use 'simple_publication_workflow'
+        self.assertEqual(self.view.get_state(title=True), 'Private')
 
     def test_context_var(self):
         pod_template = self.portal.podtemplates.get('test_template_multiple')